refactor: regroup all visualization and render in utils
